remove rate limit from BlockingQueue
Hopefully the bug I hit that prompted me to add it won't come back.
I know "hopefully the bug won't come back" sounds bad, but it was entirely
due to the overall code architecture still being up in the air, so I think
it's reasonable now that things are more orderly.