add FFTResultDistributor: does a single FFT to give to both tongue and blow
No more duplicated FFT effort. (And perhaps some duplicated PortAudio effort
too?) More subtly, it turned out that PortAudio was calling the callbacks in
chunks, like 8 in a row of one then 8 in a row of the other. It will now be
strictly alternating.

Applies to real-time usage only; the training remains unaffected (and remains
nicely distributed across available cores).