allow different sound types to use different scales. doesnt seem good.
specifically, two attempts at blow training didn't go well, because the
blow scale picker i have was giving me 0.5, which apparently wasn't a good
scale for the blowing. basically, what i have works, and this one had problems,
and i don't think it's worth trying to fix.

more fundamentally, the point (other than better flexibility among types) was to
allow the mic-up-close training to skip humming. but now that i think about it,
humming also works as a decent stand-in for people talking, so i think we want
it there as negative examples even if we are sure the user won't use it.